+++ src/Components/Endpoints/src/PublicAPI.Unshipped.txt
@@ -1,3 +1,4 @@
 #nullable enable
 Microsoft.AspNetCore.Components.Endpoints.BasePath
 Microsoft.AspNetCore.Components.Endpoints.BasePath.BasePath() -> void
+Microsoft.AspNetCore.Components.Endpoints.TempDataProviderServiceCollectionExtensions
